pyproject.toml
[project]
name = "class03-gemini-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=2.0.7",
    "python-dotenv>=1.0.1",
]
